tests/wpt/mozilla/tests/bluetooth/requestDevice/name-missing-device-from-service-filter.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.no_name_heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => assert_equals(device.name, null));
}, 'An unnamed device can be obtained by advertised service UUID.');
</script>
